# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mobile Shop POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      summary: List sales
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: customer_id
          in: query
          schema:
            type: integer
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Sale list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleListResponse"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSale
      tags: [sales]
      summary: Get sale by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
    delete:
      operationId: deleteSale
      tags: [sales]
      summary: Delete sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /sales/{id}/return:
    post:
      operationId: returnSale
      tags: [sales]
      summary: Return/refund a sale
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # PURCHASES
...
  /reports/sales:
    get:
      operationId: getSalesReport
      tags: [reports]
      summary: Get sales report
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly, yearly]
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Sales report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesReport"
  /reports/profit:
    get:
      operationId: getProfitReport
      tags: [reports]
      summary: Get profit report
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Profit report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfitReport"
  /reports/expenses:
    get:
      operationId: getExpenseReport
      tags: [reports]
      summary: Get expense report
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Expense report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseReport"
  /reports/purchases:
    get:
      operationId: getPurchaseReport
      tags: [reports]
      summary: Get purchase report
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Purchase report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseReport"
...
  /reports/best-selling:
    get:
      operationId: getBestSellingReport
      tags: [reports]
      summary: Get best selling products
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Best selling products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BestSellingItem"

  # SETTINGS
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # AUTH
...
    SaleItem:
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
        product_name:
          type: string
        barcode:
          type: ["string", "null"]
        quantity:
          type: integer
        unit_price:
          type: number
        discount:
          type: number
        total:
          type: number
      required: [id, product_id, product_name, quantity, unit_price, discount, total]
    SaleItemInput:
      type: object
      properties:
        product_id:
          type: integer
        quantity:
          type: integer
        unit_price:
          type: number
        discount:
          type: number
      required: [product_id, quantity, unit_price, discount]
    Sale:
      type: object
      properties:
        id:
          type: integer
        invoice_no:
          type: string
        customer_id:
          type: ["integer", "null"]
        customer_name:
          type: ["string", "null"]
        cashier_id:
          type: integer
        cashier_name:
          type: string
        subtotal:
          type: number
        discount_percent:
          type: number
        discount_amount:
          type: number
        tax_percent:
          type: number
        tax_amount:
          type: number
        grand_total:
          type: number
        paid_amount:
          type: number
        return_amount:
          type: number
        payment_method:
          type: string
          enum: [cash, card, easypaisa, jazzcash, credit]
        status:
          type: string
          enum: [completed, returned]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        created_at:
          type: string
      required: [id, invoice_no, cashier_id, cashier_name, subtotal, discount_percent, discount_amount, tax_percent, tax_amount, grand_total, paid_amount, return_amount, payment_method, status]
    SaleInput:
      type: object
      properties:
        customer_id:
          type: ["integer", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItemInput"
        discount_percent:
          type: number
        discount_amount:
          type: number
        tax_percent:
          type: number
        paid_amount:
          type: number
        payment_method:
          type: string
          enum: [cash, card, easypaisa, jazzcash, credit]
        notes:
          type: ["string", "null"]
      required: [items, discount_percent, discount_amount, tax_percent, paid_amount, payment_method]
    SaleListResponse:
      type: object
      properties:
        sales:
          type: array
          items:
            $ref: "#/components/schemas/Sale"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [sales, total, page, limit]

    # PURCHASE
    PurchaseItem:
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
        product_name:
          type: string
        quantity:
          type: integer
        unit_cost:
          type: number
        total:
          type: number
      required: [id, product_id, product_name, quantity, unit_cost, total]
...
    Purchase:
      type: object
      properties:
        id:
          type: integer
        invoice_no:
          type: string
        supplier_id:
          type: ["integer", "null"]
        supplier_name:
          type: ["string", "null"]
        total_amount:
          type: number
        paid_amount:
          type: number
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItem"
        created_at:
          type: string
      required: [id, invoice_no, total_amount, paid_amount]
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
        description:
          type: ["string", "null"]
        amount:
          type: number
        date:
          type: string
        created_at:
          type: string
      required: [id, category, amount, date]
...
    SalesReportItem:
      type: object
      properties:
        date:
          type: string
        sales_count:
          type: integer
        total_sales:
          type: number
        total_discount:
          type: number
        total_tax:
          type: number
        net_total:
          type: number
      required: [date, sales_count, total_sales, total_discount, total_tax, net_total]
    SalesReport:
      type: object
      properties:
        period:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/SalesReportItem"
        total_sales:
          type: number
        total_count:
          type: integer
      required: [period, data, total_sales, total_count]
    ProfitReport:
      type: object
      properties:
        total_revenue:
          type: number
        total_cost:
          type: number
        gross_profit:
          type: number
        total_expenses:
          type: number
        net_profit:
          type: number
        data:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
              revenue:
                type: number
              cost:
                type: number
              profit:
                type: number
            required: [date, revenue, cost, profit]
      required: [total_revenue, total_cost, gross_profit, total_expenses, net_profit, data]
    ExpenseReport:
      type: object
      properties:
        total_expenses:
          type: number
        by_category:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              total:
                type: number
              count:
                type: integer
            required: [category, total, count]
        data:
          type: array
          items:
            $ref: "#/components/schemas/Expense"
      required: [total_expenses, by_category, data]
    PurchaseReport:
      type: object
      properties:
        total_purchases:
          type: number
        total_count:
          type: integer
        data:
          type: array
          items:
            $ref: "#/components/schemas/Purchase"
      required: [total_purchases, total_count, data]
...
    BestSellingItem:
      type: object
      properties:
        product_id:
          type: integer
        product_name:
          type: string
        barcode:
          type: ["string", "null"]
        total_qty:
          type: integer
        total_revenue:
          type: number
      required: [product_id, product_name, total_qty, total_revenue]

    # SETTINGS
...