# lib/api-zod/src/generated/types/saleListResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Mobile Shop POS API
 * OpenAPI spec version: 0.1.0
 */
import type { Sale } from './sale';

export interface SaleListResponse {
  sales: Sale[];
  total: number;
  page: number;
  limit: number;
}